restructure, remove duplicated script code
