+++ protocol/scripts/deploys/arbitrumSepolia/templegold/99-post-deploy/99-post-deploy.ts
@@ -103,7 +103,7 @@ async function _setSpiceAuctionConfig(
         recipient: DEFAULT_SETTINGS.SPICE.AUCTION.RECIPIENT
     }
 
-    await mine(INSTANCES.TEMPLE_GOLD.SPICE_AUCTION.setAuctionConfig(config));
+    await mine(INSTANCES.TEMPLE_GOLD.SPICE_AUCTION!.setAuctionConfig(config));
 }
 
 runAsyncMain(main);
